% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_R6_kuawdc.R
\name{check_R6_kuawdc}
\alias{check_R6_kuawdc}
\title{Check compliance of an R6 class specification according to KU-AWDC standards}
\usage{
check_R6_kuawdc(Class)
}
\arguments{
\item{Class}{an R6 class to check}
}
\description{
Check compliance of an R6 class specification according to KU-AWDC standards
}
