% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatpkg-package.R
\docType{package}
\name{neatpkg-package}
\alias{neatpkg}
\alias{neatpkg-package}
\title{neatpkg: Create R Packages Quickly, Using the KU-AWDC Recommended Layout}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

}
\keyword{internal}
