% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_new.R
\name{pkg_new}
\alias{pkg_new}
\title{Create a new R package}
\usage{
pkg_new(package_name, directory = getwd())
}
\arguments{
\item{package_name}{the name of the package to create}

\item{directory}{the GitHub directory containing the package}
}
\description{
Create a new R package
}
