% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_R6.R
\name{pkg_R6}
\alias{pkg_R6}
\title{Add R6 support to a package}
\usage{
pkg_R6(package_name, directory = getwd())
}
\arguments{
\item{package_name}{name of the package (optional)}

\item{directory}{working directory containing the package}
}
\description{
Add R6 support to a package
}
